# docs-website/static/docs/api/v1/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Documentation for the traduora REST API\n\nOfficial website: https://traduora.com\nAdditional documentation: https://docs.traduora.com\nSource code: https://github.com/traduora/traduora",
    "version": "0.14.0",
    "title": "Traduora API"
  },
  "basePath": "/",
...
  "schemes": ["http", "https"],
  "securityDefinitions": {
    "oauth2": {
      "type": "oauth2",
      "flow": "password",
      "authorizationUrl": "/api/v1/auth/token",
      "tokenUrl": "/api/v1/auth/token"
    }
  },
  "paths": {
...
    "/api/v1/projects/{projectId}/imports": {
      "post": {
        "summary": "Import a translation file",
        "parameters": [
          {
            "type": "string",
            "name": "locale",
            "required": true,
            "in": "query",
            "minLength": 2,
            "maxLength": 16
          },
          {
            "type": "string",
            "name": "format",
            "required": true,
            "in": "query",
            "enum": {
              "AndroidXml": "androidxml",
              "Csv": "csv",
              "Xliff12": "xliff12",
              "JsonFlat": "jsonflat",
              "JsonNested": "jsonnested",
              "YamlFlat": "yamlflat",
              "YamlNested": "yamlnested",
              "Properties": "properties",
              "Gettext": "po",
              "Strings": "strings"
            }
          },
          {
            "type": "string",
            "name": "projectId",
            "required": true,
            "in": "path"
          },
          {
            "name": "file",
            "required": true,
            "in": "formData",
            "description": "The file to import",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "File imported",
            "schema": {
              "$ref": "#/definitions/ImportResponse"
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "402": {
            "description": "Plan limit reached"
          },
          "404": {
            "description": "No such resource found"
          }
        },
        "tags": ["Imports"],
        "consumes": ["multipart/form-data"],
        "security": [
          {
            "oauth2": []
          }
        ],
        "produces": ["application/json"]
      }
    },
...
  },
  "definitions": {
...
    "ImportTermsDTO": {
      "type": "object",
      "properties": {
        "added": {
          "type": "number"
        },
        "skipped": {
          "type": "number"
        }
      },
      "required": ["added", "skipped"]
    },
    "ImportTranslationsDTO": {
      "type": "object",
      "properties": {
        "upserted": {
          "type": "number"
        }
      },
      "required": ["upserted"]
    },
    "ImportFileDTO": {
      "type": "object",
      "properties": {
        "terms": {
          "$ref": "#/definitions/ImportTermsDTO"
        },
        "translations": {
          "$ref": "#/definitions/ImportTranslationsDTO"
        }
      },
      "required": ["terms", "translations"]
    },
    "ImportResponse": {
      "type": "object",
      "properties": {
        "data": {
          "$ref": "#/definitions/ImportFileDTO"
        }
      },
      "required": ["data"]
    },
...
  }
}